用户服务

+++ user/user-service-impl/src/main/java/com/rum/dao/UserDao.java
@@ -9,7 +9,7 @@ import org.apache.ibatis.annotations.Param;
 public interface UserDao {
     int deleteByExample(UserExample example);
 
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(User record);
 
@@ -19,7 +19,7 @@ public interface UserDao {
 
     List<User> selectByExample(UserExample example);
 
-    User selectByPrimaryKey(Long id);
+    User selectByPrimaryKey(Integer id);
 
     int updateByExampleSelective(@Param("record") User record, @Param("example") UserExample example);
